# docs/normative/v0.6/contracts/AI-05_OPENAPI.yaml
openapi: 3.1.0
info:
  title: Paquetería Local Culiacán API
  version: 0.6.0
  description: Contrato REST normativo v0.6. REST/PostgreSQL son autoridad; SignalR distribuye eventos confirmados. Seguridad
    tenant, tracking público y outbox siguen contratos ejecutables.
servers:
- url: /api/v1
security:
- oidc: []
...
paths:
...
  /settlements:
    post:
      tags:
      - Finance
      operationId: createSettlement
      description: >-
        SET-001 calculates one DRIVER settlement for a period of whole
        operational days and persists it CALCULATED in one transaction: an
        empty DRAFT header, one immutable line per eligible economic source,
        the exact bigint total, the DRAFT to CALCULATED step, the audit entry
        and the idempotency record. period_from and period_to are calendar
        dates of the configured Finance operational time zone (IANA,
        America/Mazatlan by default); the period starts at the beginning of
        period_from inclusive and ends at the beginning of the day after
        period_to exclusive, both converted to UTC. The economic sources are
        the driver's OWN and EXTERNAL dispatch assignments that bear cost
        exactly as FIN-001 counts it (ACCEPTED, ACTIVE or COMPLETED, amount
        cost_cents) whose order's latest ORD-002 DELIVERED or RETURNED status
        event occurred inside the period while the current order status still
        proves that outcome: DELIVERY for DELIVERED, CLOSED, CLAIM_OPEN or
        CLAIM_RESOLVED orders and RETURN for RETURNED orders, ordered by that
        event time. Every such line names its source as
        dispatch.assignments/{assignmentId}, the one economic identity DELIVERY
        and RETURN share. A source already held by a settlement that is not
        VOID is never included again, so a period without new sources yields
        an empty settlement with total 0. ALLY_CAPACITY work, ALLY and BUSINESS
        payees stay reserved. Only FINANCE members and PLATFORM_ADMIN members
        with a satisfied MFA challenge may operate settlements; DISPATCHER,
        DRIVER and VIEWER never may. Capability is evaluated before the
        idempotency lock, replay evidence or driver visibility, and a missing
        driver and another tenant's driver are the same uniform 404. Replaying
        the same Idempotency-Key with the same request returns the stored
        response.
      x-authorization-precedence: shape-validation-then-capability-before-persisted-state
      x-shape-validation: invalid-request-without-productive-transaction
      x-capability-protected-state:
      - idempotency_lock
      - idempotency_record
      - replay_evidence
      - driver_profile
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSettlementRequest'
      responses:
        '201':
          description: Settlement calculated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settlement'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UniformNotFound'
        '409':
          $ref: '#/components/responses/SettlementConflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    get:
      tags:
      - Finance
      operationId: listSettlements
      description: >-
        AI05-LIST-SETTLEMENTS returns one page of the settlements of the
        selected organization, newest first by created_at then id, each with
        its immutable lines exactly as getSettlement returns them. The server
        owns the bounded page size; clients cannot configure it. status keeps
        the settlements in that status; period_from and period_to keep the
        settlements whose whole period lies inside them, and both bounds may
        be used alone. Same capability as getSettlement: FINANCE members and
        PLATFORM_ADMIN members with a satisfied MFA challenge; DISPATCHER,
        DRIVER, VIEWER and every other role receive the uniform 403 before any
        settlement is read. Every query parameter appears at most once and
        parses exactly; an unknown parameter, an unknown status, a malformed
        date, period_to before period_from or a cursor that was not issued by
        this operation is INVALID_REQUEST. Another tenant's settlements are
        never listed, and a page holding a header whose total does not equal
        the exact sum of its lines fails closed as 503.
      x-authorization-precedence: shape-validation-then-capability-before-persisted-state
      x-shape-validation: invalid-request-without-productive-transaction
      x-capability-protected-state:
      - settlement
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - DRAFT
          - CALCULATED
          - APPROVED
          - PAID
          - VOID
      - name: period_from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: period_to
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: cursor
        in: query
        required: false
        schema:
          type: string
          maxLength: 128
      responses:
        '200':
          description: Cursor-paginated tenant settlements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/SettlementConflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /settlements/{settlementId}:
    get:
      tags:
      - Finance
      operationId: getSettlement
      description: >-
        SET-001 persisted settlement header and its immutable lines, ordered by
        created_at then id. Amounts are returned exactly as stored and never
        recomputed; a header whose total does not equal the exact sum of its
        lines is never shown and the read fails closed as 503. Same capability
        as every settlement operation. A missing settlement and another
        tenant's settlement are the same uniform 404.
      x-authorization-precedence: capability-before-persisted-state
      x-capability-protected-state:
      - settlement
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - $ref: '#/components/parameters/SettlementId'
      responses:
        '200':
          description: Settlement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settlement'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UniformNotFound'
        '409':
          $ref: '#/components/responses/SettlementConflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /settlements/{settlementId}/adjustments:
    post:
      tags:
      - Finance
      operationId: addSettlementAdjustment
      description: >-
        SET-001 appends one signed ADJUSTMENT line to a CALCULATED settlement
        and moves its total by exact checked bigint arithmetic; existing lines
        are never changed or replaced. The line references no order and names
        its source as platform.audit_logs/{auditId}, the append-only audit
        entry written in the same transaction, which keeps the reason. Any
        other settlement status is SETTLEMENT_STATE_CONFLICT, and an amount
        whose application would leave the bigint range is INVALID_REQUEST.
        Replaying the same Idempotency-Key with the same request returns the
        stored response.
      x-authorization-precedence: shape-validation-then-capability-before-persisted-state
      x-shape-validation: invalid-request-without-productive-transaction
      x-capability-protected-state:
      - idempotency_lock
      - idempotency_record
      - replay_evidence
      - settlement
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - $ref: '#/components/parameters/SettlementId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSettlementAdjustmentRequest'
      responses:
        '201':
          description: Adjustment added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settlement'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UniformNotFound'
        '409':
          $ref: '#/components/responses/SettlementConflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /settlements/{settlementId}/approve:
    post:
      tags:
      - Finance
      operationId: approveSettlement
      description: >-
        SET-001 approves a CALCULATED settlement, which freezes its total and
        its lines. Inside the approving transaction, after the settlement row
        lock, the current state of every included order is re-read: a
        DELIVERY whose order expects cash on delivery requires that cash
        RECONCILED for the exact expected amount (CASH_PENDING), an OPEN or
        INVESTIGATING incident on any included order blocks (INCIDENT_PENDING)
        and an included order in CLAIM_OPEN blocks (CLAIM_PENDING); RESOLVED
        and REJECTED incidents never block. A RETURN is never blocked by its
        order's COD expectation, because a returned order collected no cash.
        When several apply, the first of cash, incident and claim is reported.
        Any other settlement status is SETTLEMENT_STATE_CONFLICT. Replaying the
        same Idempotency-Key returns the stored response.
      x-authorization-precedence: shape-validation-then-capability-before-persisted-state
      x-shape-validation: invalid-request-without-productive-transaction
      x-capability-protected-state:
      - idempotency_lock
      - idempotency_record
      - replay_evidence
      - settlement
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - $ref: '#/components/parameters/SettlementId'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Settlement approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settlement'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UniformNotFound'
        '409':
          $ref: '#/components/responses/SettlementConflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/.well-known/openid-configuration
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
    OrganizationContext:
      name: X-Organization-Id
      in: header
      required: true
      description: Organización activa; debe corresponder a una membresía ACTIVE del usuario.
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateSettlementRequest:
      type: object
      additionalProperties: false
      required:
      - driver_id
      - period_from
      - period_to
      properties:
        driver_id:
          type: string
          format: uuid
          description: Driver profile of the DRIVER payee; it must belong to the selected organization.
        period_from:
          type: string
          format: date
          description: First operational day of the period, inclusive.
        period_to:
          type: string
          format: date
          description: >-
            Last operational day of the period, inclusive. It is not before
            period_from and the period covers at most 366 days.
...
    Settlement:
      type: object
      additionalProperties: false
      required:
      - id
      - payee_type
      - payee_id
      - status
      - total_cents
      - period_from
      - period_to
      - created_at
      - lines
      properties:
        id:
          type: string
          format: uuid
        payee_type:
          type: string
          enum:
          - DRIVER
        payee_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - DRAFT
          - CALCULATED
          - APPROVED
          - PAID
          - VOID
        total_cents:
          type: integer
          format: int64
          description: Always the exact sum of amount_cents over lines.
        period_from:
          type: string
          format: date
        period_to:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        lines:
          type: array
          items:
            $ref: '#/components/schemas/SettlementLine'
...